enable project sharing via URL (#189)
* enable project sharing via URL

by adding according routes to frontend

* change routes, make project selector appear only after init of ToneJS
