+++ Frontend/package.json
@@ -9,12 +9,15 @@
     "axios": "^0.19.2",
     "bootstrap": "^4.5.0",
     "file-saver": "^2.0.2",
+    "history": "^5.0.0",
     "jquery": "^3.5.1",
     "media-stream-library": "^7.0.4",
     "mp3-mediarecorder": "^4.0.3",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-hooks-global-state": "^1.0.0",
+    "react-router": "^6.0.0-beta.0",
+    "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "^3.4.1",
     "react-three-fiber": "^4.2.11",
     "react-use-gesture": "6.0.2-beta.1",
