+++ Frontend/src/components/sidemenu.js
@@ -8,8 +8,8 @@ import { ReactComponent as DeleteIcon } from '../img/delete.svg';
 import { ReactComponent as RecordIcon } from '../img/record.svg';
 import { ReactComponent as DownloadIcon } from '../img/download.svg';
 import { ReactComponent as LogoIcon } from '../img/logo.svg';
-import SaveProjectForm from './SaveProjectForm';
-import OpenProject from '../components/OpenProject';
+import SaveProject from './SaveProject';
+import OpenProjects from '../components/OpenProjects';
 import { useGlobalState } from "../GlobalState";
 
 
@@ -42,27 +42,7 @@ function SideMenu() {
 
 
 
-    //we search in the array (array includes the data from database) while typing and show the matching result
-    const findProject = () => {
-        var input, filter, ul, li, projectLi, i, txtValue;
-        input = document.getElementById('usernameProject');
-
-        filter = input.value.toUpperCase();
-        ul = document.getElementById("databaseTable");
-        li = ul.getElementsByTagName('li');
-
-        for (i = 0; i < li.length; i++) {
-            projectLi = li[i];
-            txtValue = projectLi.textContent || projectLi.innerText;
-            if (txtValue.toUpperCase().indexOf(filter) > -1) {
-                li[i].style.display = "";
-            } else {
-                li[i].style.display = "none";
-            }
-        }
-    }
-
-    //fade out side menu icons (open, save, help) when record is active
+      //fade out side menu icons (open, save, help) when record is active
     const fadeOpenSaveHelp = () => {
         let open = document.getElementById("openIcon");
         let save = document.getElementById("saveIcon");
@@ -76,15 +56,7 @@ function SideMenu() {
     }
 
 
-    //prevent the data entered in a text field from being entered in the URL line by clicking Enter
-    const preventSubmit = function (e) {
-        var key = e.charCode || e.keyCode || 0;
-        if (key === 13) {
-            e.preventDefault();
-        }
-    }
-
-
+    
     const helpProjectFunction = () => {
         setActiveHelpDialogue("canvas");
     }
@@ -154,11 +126,8 @@ function SideMenu() {
                 <>
                     <div className="saveOpenUnderlay"></div>
                     <div id="saveOverlay">
-
                         <DeleteIcon id="closeSaveOverlay" title="close overlay" onClick={() => { setSaveOverlay(false) }} />
-                        <p id="headerSave">Save your Track?</p>
-                        <SaveProjectForm />
-
+                        <SaveProject />
                     </div>
                 </>
 
@@ -170,15 +139,7 @@ function SideMenu() {
                     <div className="saveOpenUnderlay"></div>
                     <div id="openOverlay">
                         <DeleteIcon id="closeOpenOverlay" title="close overlay" onClick={() => { setOpenOverlay(false) }} />
-
-                        <p id="headerOpen">Open a Track?</p>
-                        <form>
-                            <label id="findProject">Find a project</label>
-                            <input name="usernameProject" id="usernameProject" onKeyUp={findProject} onKeyPress={preventSubmit} ></input>
-                        </form>
-                        <ul id="databaseTable">
-                            <OpenProject />
-                        </ul>
+                        <OpenProjects/>
                     </div>
                 </>
             }
